Enable assembly kernels for rv32iv
 * (Re)storing callee-saved 32-bit/64-bit registers
   is realized via a case distinction
 * The zeroing of vectors no longer relies on x0,
   the zero register, but uses either vxor
   or fcvt for scalar floating-point registers.
 * The C wrapper casts the dimensions to uintptr_t
   to ensure that the dimensions fit into the
   argument registers

Credits to Lee Killough for suggesting vxor and
the fix to get both rv32iv and rv64iv share
the interface.

+++ kernels/rviv/3/bli_cgemm_rviv_4vx4.c
@@ -37,12 +37,12 @@
 extern
 void bli_cgemm_rviv_asm_4vx4
     (
-      uint64_t               k,
+      uintptr_t              k,
       scomplex*     restrict alpha,
       scomplex*     restrict a,
       scomplex*     restrict b,
       scomplex*     restrict beta,
-      scomplex*     restrict c, uint64_t rs_c, uint64_t cs_c
+      scomplex*     restrict c, uintptr_t rs_c, uintptr_t cs_c
     );
 
 void bli_cgemm_rviv_4vx4
@@ -60,9 +60,9 @@ void bli_cgemm_rviv_4vx4
      )
 {
     // Use local copy in case dim_t has a different size than expected in the assembly kernel
-    uint64_t _k     = k;
-    uint64_t rs_c   = rs_c0;
-    uint64_t cs_c   = cs_c0;
+    uintptr_t _k     = k;
+    uintptr_t rs_c   = rs_c0;
+    uintptr_t cs_c   = cs_c0;
 
     // Extract vector-length dependent mr, nr that are fixed at configure time.
     const inc_t mr = bli_cntx_get_blksz_def_dt( BLIS_SCOMPLEX, BLIS_MR, cntx );
