+++ kernels/rviv/3/bli_zgemm_rviv_4vx4.c
@@ -38,12 +38,12 @@
 extern
 void bli_zgemm_rviv_asm_4vx4
     (
-      uint64_t               k,
+      uintptr_t              k,
       dcomplex*     restrict alpha,
       dcomplex*     restrict a,
       dcomplex*     restrict b,
       dcomplex*     restrict beta,
-      dcomplex*     restrict c, uint64_t rs_c, uint64_t cs_c
+      dcomplex*     restrict c, uintptr_t rs_c, uintptr_t cs_c
     );
 
 
@@ -62,9 +62,9 @@ void bli_zgemm_rviv_4vx4
      )
 {
     // Use local copy in case dim_t has a different size than expected in the assembly kernel
-    uint64_t _k     = k;
-    uint64_t rs_c   = rs_c0;
-    uint64_t cs_c   = cs_c0;
+    uintptr_t _k     = k;
+    uintptr_t rs_c   = rs_c0;
+    uintptr_t cs_c   = cs_c0;
 
     // Extract vector-length dependent mr, nr that are fixed at configure time.
     const inc_t mr = bli_cntx_get_blksz_def_dt( BLIS_DCOMPLEX, BLIS_MR, cntx );
